rev: revert changes to dag_1.cpp
