+++ docs/dp/code/dag/dag_1.cpp
@@ -1,10 +1,28 @@
+#include <cmath>
+#include <cstring>
 #include <iostream>
 using namespace std;
 #define MAXN (30 + 5)
+
+    
+          
+            
+    
+
+          
+          Expand Down
+          
+            
+    
+
+          
+          Expand Up
+    
+    @@ -50,20 +48,3 @@ int babylon(int n) {
+  
 #define MAXV (500 + 5)
 int d[MAXN][3];
 int x[MAXN], y[MAXN], z[MAXN];
-
 int babylon_sub(int c, int rot, int n) {
   if (d[c][rot] != -1) {
     return d[c][rot];
@@ -33,7 +51,6 @@ int babylon_sub(int c, int rot, int n) {
   }
   return d[c][rot];
 }
-
 int babylon(int n) {
   for (int i = 0; i < n; i++) {
     d[i][0] = -1;
@@ -47,4 +64,21 @@ int babylon(int n) {
     r = max(r, babylon_sub(i, 2, n) + y[i]);
   }
   return r;
+}
+
+int main() {
+  int t = 0;
+  while (true) {  // 死循环求答案
+    int n;
+    cin >> n;
+    if (n == 0) break;  // 没有砖头了就停止
+    t++;
+    for (int i = 0; i < n; i++) {
+      cin >> x[i] >> y[i] >> z[i];
+    }
+    cout << "Case " << t << ":"
+         << " maximum height = " << babylon(n);  // 递归
+    cout << endl;
+  }
+  return 0;
 }
